# rotas-openapi.yaml
openapi: 3.0.4
info:
  title: APDIS - Rotas API
  description: |-
    Representação da arquitetura REST da plataforma APDIS, desenvolvida no âmbito da unidade curricular de Projeto de Software Web, Móvel e na Nuvem.
    Neste documento será possível observar as rotas existentes, os campos necessários e respostas esperadas.
  version: 1.0.0
servers:
  - url: https://projeto.matz.pt/api
...
paths:
  /auth/isAuthorized:
    get:
      tags:
        - auth
      summary: Verifica a Sessão do Utilizador.
      description: Valida a Sessão do Utilizador Verificando se Existe.
      parameters:
        - name: clientToken
          in: header
          description: Cookie Válida do Utilizador Após o Login
          required: true
          schema:
            type: string
            default: 
      responses:
        '200':
          description: Utilizador Válido.
          content:
            application/json:
              schema:
                type: object
                properties:
                  associado:
                    type: integer
                    example: 10
                  fullname:
                    type: string
                    example: "Artur Finato"
                  email:
                    type: string
                    example: "artur.finato@example.com"
                  type:
                    type: string
                    example: "user"
        '400':
          description: Utilizador Inválido
  /auth/login:
    post:
      tags:
        - auth
      summary: Realiza a Tentativa do Pedido de Autenticação.
      description: Realiza o Pedido de Autenticação com o Número de Associado e Password.
      requestBody:
        description: Realiza a Tentativa do Pedido de Autenticação
        content:
          application/json:
            schema:
              required:
                - associado
                - password
              type: object
              properties:
                associado:
                  type: integer
                  format: int64
                  example: 10
                password:
                  type: string
                  example: password
        required: true
      responses:
        '200':
          description: Autenticação Realizada com Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Utilizador Autenticado com Sucesso."
        '400':
          description: Credenciais Inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Credenciais Inválidas."
        '500':
          description: Ocorreu um Erro
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro: ERRO"
  /auth/register:
    post:
      tags:
        - auth
      summary: Realiza a Tentativa do Pedido de Registo.
      description: Realiza a Tentativa do Pedido de Registo com o Número de Associado, E-Mail, Nome Completo, Unidade de Saúde e Cargo.
      requestBody:
        description: Realiza a Tentativa do Pedido de Registo
        content:
          application/json:
            schema:
              required:
                - associado
                - email
                - nome
                - unidsaude
                - cargo
              type: object
              properties:
                associado:
                  type: integer
                  format: int64
                  example: 10
                email:
                  type: string
                  example: "artur.finato@example.com"
                nome:
                  type: string
                  example: "Artur Finato"
                unidsaude:
                  type: string
                  example: "Cova da Beira"
                cargo:
                  type: string
                  example: "Médico"
        required: true
      responses:
        '200':
          description: Pedido de Registo Realizado com Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Pedido de Registo Enviado. \nAguarde Pela Confirmação do Responsável!"
        '400':
          description: Pedido Já Existente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Já Existe um Pedido com Esses Dados."
        '500':
          description: Credenciais Já Existentes
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Credenciais Já Existentes na Plataforma."
  /auth/password-reset:
    post:
      tags:
        - auth
      summary: Realiza a Tentativa do Pedido de Resetar a Password.
      description: Realiza a Tentativa do Pedido de Resetar a Password com o E-Mail.
      requestBody:
        description: Realiza a Tentativa do Pedido de Resetar a Password
        content:
          application/json:
            schema:
              required:
                - email
              type: object
              properties:
                email:
                  type: string
                  example: "artur.finato@example.com"
        required: true
      responses:
        '200':
          description: Pedido de Reset da Password com Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "E-Mail de Recuperação de Password Enviado."
        '400':
          description: E-Mail Não Encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "E-Mail Não Encontrado."
        '500':
          description: Ocorreu um Erro
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro: ERRO"
  /auth/password-set:
    post:
      tags:
        - auth
      summary: Realiza a Tentativa do Pedido de Atualização da Password.
      description: Realiza a Tentativa do Pedido de Atualização da Password com o Token.
      requestBody:
        description: Realiza a Tentativa do Pedido de Atualização da Password com o Token.
        content:
          application/json:
            schema:
              required:
                - token
                - password
              type: object
              properties:
                token:
                  type: string
                  example: "A8X4A6"
                password:
                  type: string
                  example: password
        required: true
      responses:
        '200':
          description: Pedido de Atualização da Password com Sucesso 
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password Atualizada com Sucesso."
        '400':
          description: E-Mail Não Encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "E-Mail Não Encontrado."
        '500':
          description: Password Fraca
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password Não é Forte o Suficiente."
  /auth/logout:
    post:
      tags:
        - auth
      summary: Limpa a Sessão do Utilizador.
      description: Limpa a Sessão do Utilizador Realizando Logout.
      parameters:
        - name: clientToken
          in: header
          description: Cookie Válida do Utilizador Após o Login
          required: true
          schema:
            type: string
            example:
      responses:
        '200':
          description: Logout Feito com Sucesso. 
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout Feito com Sucesso!"
        '500':
          description: Ocorreu um Erro
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Ocorreu um Erro: ERRO"
...